api/pyproject.toml
[project]
name = "stocking-api"
version = "0.1.0"
description = ""
authors = [
    {name = "astroshim", email = "astroshim@gmail.com"}
]
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32.3",
    "pymysql>=1.1.1",
    "gunicorn>=23.0.0",
    "pyjwt>=2.10.1",
    "botocore>=1.37.18",
    "boto3>=1.37.18",
    "cryptography>=44.0.2",
    "logging>=0.4.9.6",
    "fastapi>=0.115.12",
    "uvicorn>=0.34.0",
    "pydantic-settings>=2.8.1",
    "email-validator>=2.2.0",
    "fastapi-pagination>=0.12.34",
    "pydantic>=2.11.3",
    "python-jose>=3.4.0",
    "sqlalchemy-pagination>=0.0.2",
    "httpx>=0.28.1",
    "google-api-python-client>=2.169.0",
    "google-auth>=2.40.1",
    "google-auth-oauthlib>=1.2.2",
    "google-auth-httplib2>=0.2.0",
    "alembic>=1.16.4",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
